C4Object: Move OCF logic to a separate file
